+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ### Updated
 
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.8.0`
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.8.1`
 - `gradle -> 8.11.1`
 
 ## [0.8.0] - 2024-10-09