+++ CHANGELOG.md
@@ -14,8 +14,8 @@
 
 ### Updated
 
+- `gradle -> 8.12`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.8.1`
-- `gradle -> 8.11.1`
 
 ## [0.8.0] - 2024-10-09