+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ### Updated
 
-- `gradle -> 8.12`
+- `gradle -> 8.12.1`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.8.1`
 
 ## [0.8.0] - 2024-10-09