+++ CHANGELOG.md
@@ -14,8 +14,8 @@
 
 ### Updated
 
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.8.2`
 - `gradle -> 8.12.1`
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.8.1`
 
 ## [0.8.0] - 2024-10-09