+++ CHANGELOG.md
@@ -18,8 +18,8 @@
 
 ### Updated
 
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.14.1`
 - `gradle -> 8.14.3`
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.8.5`
 
 ## [0.8.0] - 2024-10-09