+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 
 ### Updated
 
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.14.1`
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.14.2`
 - `gradle -> 8.14.3`
 
 ## [0.8.0] - 2024-10-09